working back button reorganize components
